Forms: move empty state as component for inbox
