+++ projects/packages/forms/changelog/update-forms-inbox-empty-response-component
@@ -1,0 +1,4 @@
+Significance: minor
+Type: changed
+
+Forms: move empty state as component for inbox
